Add APP:JFrog-artifactory-oss for linix. Update OS:Alpine. Update PL:Golang.  Changes to be committed: 	modified:   app/centrifugo/Dockerfile/linux/update.exec.sh 	modified:   app/centrifugo/Dockerfile/linux/update.sh 	new file:   app/jfrog/artifactory-oss/Dockerfile/linux/update.exec.sh 	new file:   app/jfrog/artifactory-oss/Dockerfile/linux/update.sh 	modified:   app/jfrog/artifactory-oss/Dockerfile/win/update.exec.ps1 	modified:   app/jfrog/artifactory-oss/Dockerfile/win/update.ps1 	modified:   app/jfrog/artifactory-oss/README.md 	modified:   os/alpine/Dockerfile/linux/builder.sh 	modified:   os/alpine/Dockerfile/linux/run.sh 	modified:   pl/golang/Dockerfile/linux/builder.sh 	modified:   pl/golang/Dockerfile/linux/run.sh

+++ os/alpine/Dockerfile/linux/builder.sh
@@ -27,18 +27,22 @@ docker exec $uid-alpine sh -c \
 
 docker commit $uid-alpine $uid/alpine
 
-docker login
+if [ ! -z $uid ]; then
+  docker login
 
-docker tag $uid/alpine $uid/alpine:$version
-docker push $uid/alpine:$version
+  docker tag $uid/alpine $uid/alpine:$version
+  docker push $uid/alpine:$version
 
-docker tag $uid/alpine $uid/alpine
-docker push $uid/alpine
+  docker tag $uid/alpine $uid/alpine
+  docker push $uid/alpine
 
-docker stop $uid-alpine
-docker rm -v -f $uid-alpine
-docker rmi $uid/alpine:$version
-docker rmi $uid/alpine:latest
-docker rmi alpine:$version
+  docker stop $uid-alpine
+  docker rm -v -f $uid-alpine
+  docker rmi $uid/alpine:$version
+  docker rmi $uid/alpine:latest
+  docker rmi alpine:$version
 
-docker logout
+  docker logout
+fi
+
+echo Success
